backend/requirements.txt
pandas
matplotlib
seaborn
qdrant-client
sentence-transformers
llama-cpp-python
huggingface_hub
rich
langchain
langchain-community
langchain-huggingface
langchain-qdrant
paho-mqtt
langgraph
fastapi
uvicorn
sse-starlette
python-multipart
